put algs back in game.js
